Emacs 23.4.1 returns symbols from `vc-backend`, not strings.
`(vc-backend)` is the symbol `GIT` in some versions of Emacs.  Rather than
just comparing it with the fixed case string, case-fold it for comparison.

This makes the detection of the back-end much more robust - extracting the
logic to determine if the buffer is managed by git, then using that in all the
hooks.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>