Invoke `git-wip` through git magic, not directly.
In modern git installations, most of the `git-whatever` commands are hidden
well away from the user - and so are the support components like the shell
setup libraries and so forth.

That means that invoking `git-wip` directly fails, but `git wip` will do
exactly the right thing.  Since the cost of invoking it that way is low, we
might as well...

Signed-off-by: Daniel Pittman <daniel@rimspace.net>