Allow gpg commit signing to be disabled
With 'commit.gpgSign = true', git will sign all commits, including
those made via commit-tree.  This is good -- except that when git-wip is
invoked from an editor this often results in a mess on one's screen from
a "You need a passphrase to unlock your secret key..." message.

Unfortunately, git does not appear to have any way to pass "--batch" or
other options to gpg.  It does, however, allow commands to be called
with '--no-gpg-sign' in order to override the global "sign all the
commits!" configuration.

This commit allows one to disable commit signing for WIP commits by
adding a '--no-gpg-sign' option to git-wip.  It also adds a check for
g:git_wip_disable_signing to the vim plugin; if true it causes git-wip
to be called with the new --no-gpg-sign flag.